drafted genome partitioning by sliding windows with the capacity for overlapping windows with filtering for windows that are complete subsets of larger windows along the terminus of chromosomes or scaffolds

+++ src/tables/tajima_d.rs
@@ -9,8 +9,8 @@ use std::time::{SystemTime, UNIX_EPOCH};
 pub fn tajima_d(
     genotypes_and_phenotypes: &GenotypesAndPhenotypes,
     pool_sizes: &Vec<f64>,
-    window_size_bp: &usize,
-    min_loci_per_window: &usize,
+    window_size_bp: &u64,
+    min_loci_per_window: &u64,
     fname_input: &String,
     fname_output: &String,
 ) -> io::Result<String> {
